Ease Your Day
An application that allows you to keep track of your tasks, upcoming deadlines, reminders and upcoming events! Helps an individual stay organized, productive and up to date. The website allows you to add tasks, delete tasks, make changes to a task and set up more important tasks needed to be done. With this, you feel an ease in your day!